p2p: make peer scoring test more resilient (#6322)
